Fix CFG generator
Fixed the CFG generator as it did not create items on new lines plus it had the wrong variable for the output dir. The output dir line also lacked a slash for the file name. Fixes also include deleting all of the CFG files in the selected/default ARF directory present beforehand so that compatibility issues do not arise.

Fixed a bug with the nbr_path variable not being given to the call command correctly.

Initial testing has provided evidence that the call command for the NBR player to execute the conversion is not working. Testing various ways to correct this.