GUI Part 2
Renamed Auto Convert to Auto_Convert so that it could be imported into a
diffrent script. Added the ability to be run as a stand alone program or
as a library to the main program. Part 2 of the GUI work has been
completed. Working GUI for basic conversions. No options or progress
bars are available. Program will freeze while converting.