Add Classes, Remove globalization of varibles
Added a class to ease the creation of GUI window systems. Switched
init_script to a class to reduce the amount of global varibles needed. A
sideaffect of the init_script function removal was the reduction of
needed functions. Script size was greatly reduced. Initial help system
function, intorduced. The only global varibles needed now are for the
window systems, so that windows can be controled from diffrent
functions.