Dictionary update - Console Version
Dictionary based vars now instead of standard vars. Not really needed.
Converted to dictionary based setup to be more in sync with the GUI
version (easier to share code back and forth). Updated the [UI] section
code to use ints instead of bools.
Updated git ignore to ignore ARF files in teh working directory and to
ignore the default Converted folder (useful if you are testing the
scripts with the default settings)