Lambda Example
Implemented lambda test/example as a button. The purpose for this code
is to demonstrate how to pass arguments via a button press without
creating extra code/functions to do so.

Also updated class name for pep8 compliency, moved another function out
of Render_Window as Render_Window's scope does not handle window closing
commands, that is for the program implementing Render_Window to handle.