Rename file
Rename the file to Convert instead of Covert.

Add initial parameter support, update the help doc, start adding tests to the system requirements testing function.

Add a requires statement to force it to only run on Windows by forcing Windows Powershell instead of powershell core. This will be changed in the future to allow execution from modern versions of powershell instead of version 5.